+++ gc/gc-repository-jdbc/src/intTest/java/org/projectnessie/gc/contents/jdbc/ITMySQLPersistenceSpi.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Dremio
+ * Copyright (C) 2024 Dremio
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,22 +15,20 @@
  */
 package org.projectnessie.gc.contents.jdbc;
 
+import static org.projectnessie.gc.contents.jdbc.ITPostgresPersistenceSpi.dockerImage;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.projectnessie.nessie.testing.containerspec.ContainerSpecHelper;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.utility.DockerImageName;
+import org.testcontainers.containers.MySQLContainer;
 
-public class ITPostgresPersistenceSpi extends AbstractJdbcPersistenceSpi {
+public class ITMySQLPersistenceSpi extends AbstractJdbcPersistenceSpi {
 
-  private static PostgreSQLContainer<?> container;
+  private static MySQLContainer<?> container;
 
   @BeforeAll
   static void createDataSource() throws Exception {
-
-    container = new PostgreSQLContainer<>(dockerImage("postgres"));
+    container = new MariaDBDriverMySQLContainer(dockerImage("mysql"));
     container.start();
-
     initDataSource(container.getJdbcUrl());
   }
 
@@ -41,12 +39,21 @@ public class ITPostgresPersistenceSpi extends AbstractJdbcPersistenceSpi {
     }
   }
 
-  public static DockerImageName dockerImage(String dbName) {
-    return ContainerSpecHelper.builder()
-        .name(dbName)
-        .containerClass(ITPostgresPersistenceSpi.class)
-        .build()
-        .dockerImageName(null)
-        .asCompatibleSubstituteFor("postgres");
+  private static class MariaDBDriverMySQLContainer
+      extends MySQLContainer<MariaDBDriverMySQLContainer> {
+
+    MariaDBDriverMySQLContainer(String dockerImage) {
+      super(dockerImage);
+    }
+
+    @Override
+    public String getDriverClassName() {
+      return "org.mariadb.jdbc.Driver";
+    }
+
+    @Override
+    public String getJdbcUrl() {
+      return super.getJdbcUrl().replace("jdbc:mysql", "jdbc:mariadb");
+    }
   }
 }