+++ gc/gc-repository-jdbc/src/main/java/org/projectnessie/gc/contents/jdbc/SqlDmlDdl.java
@@ -71,8 +71,7 @@ final class SqlDmlDdl {
   @Language("SQL")
   static final String INSERT_FILE_DELETIONS =
       "INSERT INTO gc_file_deletions \n"
-          + "    (live_set_id, base_uri, path_uri, modification_timestamp) VALUES (?, ?, ?, ?) \n"
-          + "    ON CONFLICT DO NOTHING";
+          + "    (live_set_id, base_uri, path_uri, modification_timestamp) VALUES (?, ?, ?, ?)";
 
   @Language("SQL")
   static final String SELECT_FILE_DELETIONS =
@@ -109,8 +108,7 @@ final class SqlDmlDdl {
   @Language("SQL")
   static final String INSERT_CONTENT_LOCATION =
       "INSERT INTO gc_live_set_content_locations \n"
-          + "    (live_set_id, content_id, base_location) VALUES (?, ?, ?) \n"
-          + "    ON CONFLICT DO NOTHING";
+          + "    (live_set_id, content_id, base_location) VALUES (?, ?, ?)";
 
   @Language("SQL")
   static final String SELECT_CONTENT_LOCATION =
@@ -159,8 +157,7 @@ final class SqlDmlDdl {
   static final String ADD_CONTENT =
       "INSERT INTO gc_live_set_contents \n"
           + "    (live_set_id, content_id, commit_id, content_key, content_type, metadata_location, snapshot_id) \n"
-          + "    VALUES (?, ?, ?, ?, ?, ?, ?) \n"
-          + "    ON CONFLICT DO NOTHING";
+          + "    VALUES (?, ?, ?, ?, ?, ?, ?)";
 
   @Language("SQL")
   static final String SELECT_CONTENT_REFERENCES =